+++ model.go
@@ -146,8 +146,9 @@ type TotalVisitorStats struct {
 // VisitorStats is the result type for visitor statistics.
 type VisitorStats struct {
 	Day        time.Time `json:"day"`
-	Week       int       `json:"week"`
-	Year       int       `json:"year"`
+	Week       time.Time `json:"week"`
+	Month      time.Time `json:"month"`
+	Year       time.Time `json:"year"`
 	Visitors   int       `json:"visitors"`
 	Views      int       `json:"views"`
 	Sessions   int       `json:"sessions"`
@@ -263,8 +264,9 @@ type PlatformStats struct {
 // TimeSpentStats is the result type for average time spent statistics (sessions, time on page).
 type TimeSpentStats struct {
 	Day                     time.Time `json:"day"`
-	Week                    int       `json:"week"`
-	Year                    int       `json:"year"`
+	Week                    time.Time `json:"week"`
+	Month                   time.Time `json:"month"`
+	Year                    time.Time `json:"year"`
 	Path                    string    `json:"path"`
 	Title                   string    `json:"title"`
 	AverageTimeSpentSeconds int       `db:"average_time_spent_seconds" json:"average_time_spent_seconds"`