+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 * added script to update and clean up the User-Agent blacklist
 * added flagging bots based on how quickly page views are being sent
 * added options to `Analyzer`
-* added grouping results by week and year and new fields `Week` and `Year` to `VisitorStats` and `TimeSpentStats`
+* added grouping results by week, month, and year and new fields `Week`, `Month`, and `Year` to `VisitorStats` and `TimeSpentStats`
 * calculating the growth for today will now take the time into account, and not compare to the full past day
 * optimized JS scripts and added page filtering and option to disable the script using localStorage
 * removed `Filter.Day` and `Filter.Start` and added `Filter.IncludeTime` instead, allowing to filter everything by date and time